+++ nilearn/glm/tests/test_second_level.py
@@ -38,9 +38,7 @@ from nilearn.glm.second_level.second_level import (
 )
 from nilearn.image import concat_imgs, get_data, new_img_like, smooth_img
 from nilearn.maskers import NiftiMasker
-from nilearn.maskers._utils import (
-    concatenate_surface_images,
-)
+from nilearn.surface import concat_imgs as surf_concat_imgs
 from nilearn.surface._testing import assert_surface_image_equal
 
 if is_matplotlib_installed():
@@ -1487,7 +1485,7 @@ def test_second_level_input_as_surface_image_3d_same_as_list_2d(surf_img_1d):
     model.fit(second_level_input, design_matrix=design_matrix)
     result_2d = model.compute_contrast()
 
-    second_level_input_3d = concatenate_surface_images(second_level_input)
+    second_level_input_3d = surf_concat_imgs(second_level_input)
     model.fit(second_level_input_3d, design_matrix=design_matrix)
     result_3d = model.compute_contrast()
