+++ nilearn/maskers/surface_labels_masker.py
@@ -11,13 +11,15 @@ from nilearn._utils.cache_mixin import cache
 from nilearn._utils.class_inspect import get_params
 from nilearn._utils.helpers import is_matplotlib_installed
 from nilearn.maskers._utils import (
-    check_same_n_vertices,
-    compute_mean_surface_image,
-    concatenate_surface_images,
     get_min_max_surface_image,
 )
 from nilearn.maskers.base_masker import _BaseSurfaceMasker
-from nilearn.surface import SurfaceImage
+from nilearn.surface import (
+    SurfaceImage,
+    check_same_n_vertices,
+    concat_imgs,
+    mean_img,
+)
 
 
 def _apply_mask(labels_masker, mask_data, labels_data):
@@ -321,7 +323,7 @@ class SurfaceLabelsMasker(_BaseSurfaceMasker):
         # to be able to concatenate it
         if not isinstance(img, list):
             img = [img]
-        img = concatenate_surface_images(img)
+        img = concat_imgs(img)
         check_same_n_vertices(self.labels_img.mesh, img.mesh)
         # concatenate data over hemispheres
         img_data = np.concatenate(list(img.data.parts.values()), axis=0)
@@ -521,7 +523,7 @@ class SurfaceLabelsMasker(_BaseSurfaceMasker):
 
         img = self._reporting_data["images"]
         if img:
-            img = compute_mean_surface_image(img)
+            img = mean_img(img)
             vmin, vmax = get_min_max_surface_image(img)
 
         # TODO: possibly allow to generate a report with other views
