+++ nilearn/maskers/surface_masker.py
@@ -13,13 +13,15 @@ from nilearn._utils.cache_mixin import cache
 from nilearn._utils.class_inspect import get_params
 from nilearn._utils.helpers import is_matplotlib_installed
 from nilearn.maskers._utils import (
-    check_same_n_vertices,
-    compute_mean_surface_image,
-    concatenate_surface_images,
     get_min_max_surface_image,
 )
 from nilearn.maskers.base_masker import _BaseSurfaceMasker
-from nilearn.surface import SurfaceImage
+from nilearn.surface import (
+    SurfaceImage,
+    check_same_n_vertices,
+    concat_imgs,
+    mean_img,
+)
 
 
 @fill_doc
@@ -165,7 +167,7 @@ class SurfaceMasker(_BaseSurfaceMasker):
 
         if not isinstance(img, list):
             img = [img]
-        img = concatenate_surface_images(img)
+        img = concat_imgs(img)
 
         # TODO: don't store a full array of 1 to mean "no masking"; use some
         # sentinel value
@@ -276,7 +278,7 @@ class SurfaceMasker(_BaseSurfaceMasker):
 
         if not isinstance(img, list):
             img = [img]
-        img = concatenate_surface_images(img)
+        img = concat_imgs(img)
 
         check_same_n_vertices(self.mask_img_.mesh, img.mesh)
 
@@ -467,7 +469,7 @@ class SurfaceMasker(_BaseSurfaceMasker):
         vmax = None
         if self._reporting_data["images"]:
             background_data = self._reporting_data["images"]
-            background_data = compute_mean_surface_image(background_data)
+            background_data = mean_img(background_data)
             vmin, vmax = get_min_max_surface_image(background_data)
 
         views = ["lateral", "medial"]
