+++ nilearn/surface/surface.py
@@ -1519,6 +1519,31 @@ def _check_data_and_mesh_compat(mesh, data):
             )
 
 
+def check_same_n_vertices(mesh_1, mesh_2):
+    """Check that 2 PolyMesh have the same keys and that n vertices match.
+
+    Parameters
+    ----------
+    mesh_1: PolyMesh
+
+    mesh_2: PolyMesh
+    """
+    keys_1, keys_2 = set(mesh_1.parts.keys()), set(mesh_2.parts.keys())
+    if keys_1 != keys_2:
+        diff = keys_1.symmetric_difference(keys_2)
+        raise ValueError(
+            f"Meshes do not have the same keys. Offending keys: {diff}"
+        )
+    for key in keys_1:
+        if mesh_1.parts[key].n_vertices != mesh_2.parts[key].n_vertices:
+            raise ValueError(
+                f"Number of vertices do not match for '{key}'."
+                "number of vertices in mesh_1: "
+                f"{mesh_1.parts[key].n_vertices}; "
+                f"in mesh_2: {mesh_2.parts[key].n_vertices}"
+            )
+
+
 def _mesh_to_gifti(coordinates, faces, gifti_file):
     """Write surface mesh to gifti file on disk.
 
@@ -1753,3 +1778,61 @@ class SurfaceImage:
         data = PolyData(left=texture_left, right=texture_right)
 
         return cls(mesh=mesh, data=data)
+
+
+def concat_imgs(imgs):
+    """Concatenate the data of a list or tuple of SurfaceImages.
+
+    Assumes all images have same meshes.
+
+    Parameters
+    ----------
+    imgs : :obj:`list` or :obj:`tuple` of SurfaceImage object
+
+    Returns
+    -------
+    SurfaceImage object
+    """
+    if not isinstance(imgs, (tuple, list)) or any(
+        not isinstance(x, SurfaceImage) for x in imgs
+    ):
+        raise TypeError(
+            "'imgs' must be a list or a tuple of SurfaceImage instances."
+        )
+
+    if len(imgs) == 1:
+        return imgs[0]
+
+    for i, img in enumerate(imgs):
+        check_same_n_vertices(img.mesh, imgs[0].mesh)
+        imgs[i] = at_least_2d(img)
+
+    output_data = {}
+    for part in imgs[0].data.parts:
+        tmp = [img.data.parts[part] for img in imgs]
+        output_data[part] = np.concatenate(tmp, axis=1)
+
+    output = SurfaceImage(mesh=imgs[0].mesh, data=output_data)
+
+    return output
+
+
+def mean_img(img):
+    """Compute mean of SurfaceImage over time points (for 'time series').
+
+    Parameters
+    ----------
+    img : SurfaceImage
+
+    Returns
+    -------
+    SurfaceImage
+    """
+    if len(img.shape) < 2 or img.shape[1] < 2:
+        return img
+
+    data = {
+        part: np.mean(value, axis=1).astype(float)
+        for part, value in img.data.parts.items()
+    }
+    return SurfaceImage(mesh=img.mesh, data=data)
